Widen the last two reverb locals the tank fix missed
The pan-delay branch read panDelayFeedback and delayGain into float locals,
narrowing two fields the tank fix had already widened.  Every other local in
that file was converted with it; these two sit inside a switch case and were
missed.

No measured change -- the gs_reverb tests that exercise this branch do not move
-- but the narrowing was real and in a feedback path.

+++ spessasynth_core/src/synthesizer/audio_engine/effects/reverb/reverb.c
@@ -427,7 +427,7 @@ void ss_reverb_process(SS_Reverb *reverb,
 			}
 
 			// Mix right into left
-			const float fb = reverb->panDelayFeedback;
+			const double fb = reverb->panDelayFeedback;
 			float *delayLeftInput = reverb->delayLeftInput;
 			float *delayLeftOutput = reverb->delayLeftOutput;
 			float *delayRightOutput = reverb->delayRightOutput;
@@ -439,7 +439,7 @@ void ss_reverb_process(SS_Reverb *reverb,
 			// Process right
 			ss_delay_line_process(reverb->delayRight, delayLeftOutput, delayRightOutput, sample_count);
 			// Mix
-			const float g = reverb->delayGain;
+			const double g = reverb->delayGain;
 			for(i = 0; i < sample_count; i++) {
 				*outputL++ += delayLeftOutput[i] * g;
 				*outputR++ += delayRightOutput[i] * g;
